Add example with interrupts

+++ examples/ToggleLedOnDoubleClick/ToggleLedOnDoubleClick.ino
@@ -17,8 +17,8 @@
 #include <ObjectButton.h>
 #include <interfaces/IOnDoubleClickListener.h>
 
-constexpr static int INPUT_PIN = A1;
-constexpr static int LED_PIN = LED_BUILTIN;
+constexpr static byte INPUT_PIN = A1;
+constexpr static byte LED_PIN = LED_BUILTIN;
 
 class ToggleLedOnDoubleClick : private virtual IOnDoubleClickListener {
 public:
@@ -31,7 +31,7 @@ private:
     void onDoubleClick(ObjectButton &button) override;
 
     ObjectButton button = ObjectButton(INPUT_PIN, /* activeLow */ true);
-    int ledState = LOW;
+    byte ledState = LOW;
 };
 
 void ToggleLedOnDoubleClick::onDoubleClick(ObjectButton &button) {
@@ -44,7 +44,7 @@ void ToggleLedOnDoubleClick::onDoubleClick(ObjectButton &button) {
 }
 
 void ToggleLedOnDoubleClick::init() {
-    // Setup the Serial port. see http://arduino.cc/en/Serial/IfSerial
+    // Setup the Serial port. See http://arduino.cc/en/Serial/IfSerial
     Serial.begin(9600);
     while (!Serial) {
         ; // wait for serial port to connect. Needed for Leonardo only
